Add files via upload
A Java program that creates math flashcards for elementary grade students. The user will enter his/her name, the type of problems to be worked (A, S, M, or D), the range of the factors to be used in the problems, and the number of problems to work. The system will present problems, evaluate user responses to the problems, score the problems, and provide statistics about the session at the end